+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -283,4 +283,8 @@ export default {
     'es.math.f16round',
     'es.regexp.escape',
   ],
+  3.43: [
+    'esnext.iterator.chunks',
+    'esnext.iterator.windows',
+  ],
 };