+++ tests/compat/tests.js
@@ -1824,6 +1824,9 @@ GLOBAL.tests = {
   'esnext.function.metadata': function () {
     return Function.prototype[Symbol.metadata] === null;
   },
+  'esnext.iterator.chunks': function () {
+    return Iterator.prototype.chunks;
+  },
   'esnext.iterator.concat': function () {
     return Iterator.concat;
   },
@@ -1835,6 +1838,9 @@ GLOBAL.tests = {
   },
   'esnext.iterator.to-async': function () {
     return Iterator.prototype.toAsync;
+  },
+  'esnext.iterator.windows': function () {
+    return Iterator.prototype.windows;
   },
   'esnext.json.is-raw-json': NATIVE_RAW_JSON,
   'esnext.json.parse': function () {