+++ tests/eslint/eslint.config.js
@@ -1256,6 +1256,8 @@ const forbidCompletelyNonExistentBuiltIns = {
     // TODO: drop from `core-js@4`
     'asIndexedPairs',
     'indexed',
+    'chunks',
+    'windows',
   ] }],
   'es/no-nonstandard-json-properties': [ERROR, { allow: [
     'isRawJSON',